Install gem bundit and authorization for edit delete for the user who has created

+++ app/controllers/words_controller.rb
@@ -20,10 +20,12 @@ class WordsController < ApplicationController
     end
     
     def edit
+        authorize @word
     end
 
     def update
         @word = Word.find(params[:id])
+        authorize @word
         if @word.update(word_params)
             redirect_to(word_path(@word))
         else
@@ -35,6 +37,7 @@ class WordsController < ApplicationController
     end
 
     def destroy
+        authorize @word
         if @word.destroy
             redirect_to(words_path)
         else
